Refactor DashboardPostDetailActivity to separate adapter and models
Extracted `PostDetailAdapter` and `PostDetailItem` into their own files.
Made internal helper functions available for the adapter to use.
