+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/DashboardPostDetailActivity.kt
@@ -72,7 +72,7 @@ import org.ole.planet.myplanet.lite.profile.UserProfileDatabase
 import org.ole.planet.myplanet.lite.util.MarkdownUtils
 import org.ole.planet.myplanet.lite.util.SecurePreferencesProvider
 
-private fun transformCommentMarkdownForDisplay(markdown: String): String {
+internal fun transformCommentMarkdownForDisplay(markdown: String): String {
     return markdown.replace("\n", "  \n")
 }
 
@@ -1546,340 +1546,6 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         editVoiceLauncher.launch(intent)
     }
 
-    private class PostDetailAdapter(
-        private val markwon: Markwon,
-        private val avatarBinder: (ImageView, String?, Boolean) -> Unit,
-        private val imageBinder: (ImageView, String) -> Unit,
-        private val onImageClicked: (List<String>, Int) -> Unit,
-        private val onDeleteClicked: () -> Unit,
-        private val onShareClicked: (PostDetailItem.Header) -> Unit,
-        private val onEditClicked: (PostDetailItem.Header) -> Unit,
-        private val onReplyClicked: () -> Unit,
-        private val onCommentEditClicked: (PostDetailItem.Comment) -> Unit,
-        private val onCommentDeleteClicked: (PostDetailItem.Comment) -> Unit
-    ) : androidx.recyclerview.widget.ListAdapter<PostDetailItem, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
-
-        override fun getItemViewType(position: Int): Int {
-            return when (getItem(position)) {
-                is PostDetailItem.Header -> VIEW_TYPE_HEADER
-                is PostDetailItem.Comment -> VIEW_TYPE_COMMENT
-            }
-        }
-
-        override fun onCreateViewHolder(parent: android.view.ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-            val inflater = android.view.LayoutInflater.from(parent.context)
-            return when (viewType) {
-                VIEW_TYPE_HEADER -> {
-                    val view = inflater.inflate(R.layout.item_dashboard_post_detail_header, parent, false)
-                    HeaderViewHolder(
-                        view,
-                        markwon,
-                        avatarBinder,
-                        imageBinder,
-                        onImageClicked,
-                        onDeleteClicked,
-                        onShareClicked,
-                        onEditClicked,
-                        onReplyClicked
-                    )
-                }
-                else -> {
-                    val view = inflater.inflate(R.layout.item_dashboard_comment, parent, false)
-                    CommentViewHolder(
-                        view,
-                        markwon,
-                        avatarBinder,
-                        imageBinder,
-                        onImageClicked,
-                        onCommentEditClicked,
-                        onCommentDeleteClicked
-                    )
-                }
-            }
-        }
-
-        override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-            val item = getItem(position)
-            when {
-                holder is HeaderViewHolder && item is PostDetailItem.Header -> holder.bind(item)
-                holder is CommentViewHolder && item is PostDetailItem.Comment -> {
-                    val isLast = position == itemCount - 1
-                    holder.bind(item, isLast)
-                }
-            }
-        }
-
-        private class HeaderViewHolder(
-            view: View,
-            private val markwon: Markwon,
-            private val avatarBinder: (ImageView, String?, Boolean) -> Unit,
-            private val imageBinder: (ImageView, String) -> Unit,
-            private val onImageClicked: (List<String>, Int) -> Unit,
-            private val onDeleteClicked: () -> Unit,
-            private val onShareClicked: (PostDetailItem.Header) -> Unit,
-            private val onEditClicked: (PostDetailItem.Header) -> Unit,
-            private val onReplyClicked: () -> Unit
-        ) : RecyclerView.ViewHolder(view) {
-
-            private val authorView: TextView = view.findViewById(R.id.postDetailAuthor)
-            private val metadataView: TextView = view.findViewById(R.id.postDetailMetadata)
-            private val bodyView: TextView = view.findViewById(R.id.postDetailBody)
-            private val avatarView: ImageView = view.findViewById(R.id.postDetailAvatar)
-            private val imagesContainer: LinearLayout = view.findViewById(R.id.postDetailImagesContainer)
-            private val commentsLabel: TextView = view.findViewById(R.id.postDetailCommentsLabel)
-            private val commentsEmpty: TextView = view.findViewById(R.id.postDetailCommentsEmpty)
-            private val dividerView: View = view.findViewById(R.id.postDetailDivider)
-            private val overflowMenu: View = view.findViewById(R.id.postDetailOverflowMenu)
-
-            fun bind(item: PostDetailItem.Header) {
-                authorView.text = item.author
-                val relativeTime = formatRelativeTime(itemView.context, item.timestamp)
-                metadataView.text = buildMetadata(item.username, relativeTime)
-                if (item.message.isNullOrBlank()) {
-                    bodyView.isVisible = false
-                    bodyView.text = ""
-                } else {
-                    bodyView.isVisible = true
-                    val renderedMessage = transformCommentMarkdownForDisplay(item.message)
-                    markwon.setMarkdown(bodyView, renderedMessage)
-                }
-                avatarBinder(avatarView, item.username, item.hasAvatar)
-                bindImages(item)
-                val count = item.commentCount.coerceAtLeast(0)
-                commentsLabel.isVisible = true
-                commentsEmpty.isVisible = !item.isLoadingComments && count == 0
-                dividerView.isVisible = count > 0
-                bindActions(item)
-            }
-
-            @SuppressLint("RestrictedApi")
-            private fun bindActions(item: PostDetailItem.Header) {
-                val hasActions = item.canReply || item.canEdit || item.canDelete || item.canShare
-                overflowMenu.isVisible = hasActions
-                if (!hasActions) {
-                    overflowMenu.setOnClickListener(null)
-                    return
-                }
-
-                overflowMenu.setOnClickListener {
-                    val themedContext = ContextThemeWrapper(itemView.context, R.style.Widget_MyPlanet_PopupMenu)
-                    val popup = PopupMenu(themedContext, overflowMenu)
-                    popup.menuInflater.inflate(R.menu.menu_dashboard_post_actions, popup.menu)
-                    popup.menu.findItem(R.id.action_reply).isVisible = item.canReply
-                    popup.menu.findItem(R.id.action_edit).isVisible = item.canEdit
-                    popup.menu.findItem(R.id.action_delete).isVisible = item.canDelete
-                    popup.menu.findItem(R.id.action_share).isVisible = item.canShare
-                    if (popup.menu is MenuBuilder) {
-                        (popup.menu as MenuBuilder).setOptionalIconsVisible(true)
-                    }
-
-                    popup.setOnMenuItemClickListener { menuItem ->
-                        when (menuItem.itemId) {
-                            R.id.action_reply -> {
-                                onReplyClicked()
-                                true
-                            }
-
-                            R.id.action_edit -> {
-                                onEditClicked(item)
-                                true
-                            }
-
-                            R.id.action_delete -> {
-                                onDeleteClicked()
-                                true
-                            }
-
-                            R.id.action_share -> {
-                                onShareClicked(item)
-                                true
-                            }
-
-                            else -> false
-                        }
-                    }
-
-                    popup.show()
-                }
-            }
-
-            private fun bindImages(item: PostDetailItem.Header) {
-                val imagePaths = item.imagePaths
-                if (imagePaths.isEmpty()) {
-                    imagesContainer.isVisible = false
-                    imagesContainer.removeAllViews()
-                    return
-                }
-                imagesContainer.isVisible = true
-                imagesContainer.removeAllViews()
-                val context = imagesContainer.context
-                val spacing = context.resources.getDimensionPixelSize(R.dimen.dashboard_post_image_spacing)
-                imagePaths.forEachIndexed { index, path ->
-                    val imageView = AppCompatImageView(context)
-                    val params = LinearLayout.LayoutParams(
-                        android.view.ViewGroup.LayoutParams.MATCH_PARENT,
-                        android.view.ViewGroup.LayoutParams.WRAP_CONTENT
-                    )
-                    if (index > 0) {
-                        params.topMargin = spacing
-                    }
-                    imageView.layoutParams = params
-                    imageView.adjustViewBounds = true
-                    imageView.scaleType = ImageView.ScaleType.FIT_CENTER
-                    imageView.setBackgroundResource(R.drawable.dashboard_post_image_placeholder)
-                    imageView.contentDescription = context.getString(R.string.dashboard_post_image_content_description)
-                    imagesContainer.addView(imageView)
-                    imageBinder(imageView, path)
-                    imageView.setOnClickListener { onImageClicked(item.imagePaths, index) }
-                }
-            }
-            companion object {
-            }
-        }
-
-        private class CommentViewHolder(
-            view: View,
-            private val markwon: Markwon,
-            private val avatarBinder: (ImageView, String?, Boolean) -> Unit,
-            private val imageBinder: (ImageView, String) -> Unit,
-            private val onImageClicked: (List<String>, Int) -> Unit,
-            private val onEditClicked: (PostDetailItem.Comment) -> Unit,
-            private val onDeleteClicked: (PostDetailItem.Comment) -> Unit
-        ) : RecyclerView.ViewHolder(view) {
-
-            private val authorView: TextView = view.findViewById(R.id.commentAuthor)
-            private val metadataView: TextView = view.findViewById(R.id.commentMetadata)
-            private val bodyView: TextView = view.findViewById(R.id.commentBody)
-            private val avatarView: ImageView = view.findViewById(R.id.commentAvatar)
-            private val imagesContainer: LinearLayout = view.findViewById(R.id.commentImagesContainer)
-            private val dividerView: View = view.findViewById(R.id.commentDivider)
-            private val overflowMenu: View = view.findViewById(R.id.commentOverflowMenu)
-
-            fun bind(item: PostDetailItem.Comment, isLast: Boolean) {
-                authorView.text = item.author
-                val relativeTime = formatRelativeTime(itemView.context, item.timestamp)
-                metadataView.text = buildMetadata(item.username, relativeTime)
-                if (item.message.isNullOrBlank()) {
-                    bodyView.isVisible = false
-                    bodyView.text = ""
-                } else {
-                    bodyView.isVisible = true
-                    val renderedMessage = transformCommentMarkdownForDisplay(item.message)
-                    markwon.setMarkdown(bodyView, renderedMessage)
-                }
-                if (item.imagePaths.isEmpty()) {
-                    imagesContainer.isVisible = false
-                    imagesContainer.removeAllViews()
-                } else {
-                    val context = imagesContainer.context
-                    val spacing = context.resources.getDimensionPixelSize(R.dimen.dashboard_post_image_spacing)
-                    imagesContainer.removeAllViews()
-                    item.imagePaths.forEachIndexed { index, path ->
-                        val imageView = AppCompatImageView(context)
-                        val params = LinearLayout.LayoutParams(
-                            android.view.ViewGroup.LayoutParams.MATCH_PARENT,
-                            android.view.ViewGroup.LayoutParams.WRAP_CONTENT
-                        )
-                        if (index > 0) {
-                            params.topMargin = spacing
-                        }
-                        imageView.layoutParams = params
-                        imageView.adjustViewBounds = true
-                        imageView.scaleType = ImageView.ScaleType.FIT_CENTER
-                        imageView.setBackgroundResource(R.drawable.dashboard_post_image_placeholder)
-                        imageView.contentDescription = context.getString(R.string.dashboard_post_image_content_description)
-                        imagesContainer.addView(imageView)
-                        imageBinder(imageView, path)
-                        imageView.setOnClickListener { onImageClicked(item.imagePaths, index) }
-                    }
-                    imagesContainer.isVisible = true
-                }
-                avatarBinder(avatarView, item.username, item.hasAvatar)
-                bindActions(item)
-                dividerView.isVisible = !isLast
-            }
-
-            @SuppressLint("RestrictedApi")
-            private fun bindActions(item: PostDetailItem.Comment) {
-                val hasActions = item.canEdit || item.canDelete
-                overflowMenu.isVisible = hasActions
-                if (!hasActions) {
-                    overflowMenu.setOnClickListener(null)
-                    return
-                }
-                overflowMenu.setOnClickListener {
-                    val themedContext = ContextThemeWrapper(itemView.context, R.style.Widget_MyPlanet_PopupMenu)
-                    val popup = PopupMenu(themedContext, overflowMenu)
-                    popup.menuInflater.inflate(R.menu.menu_dashboard_comment_actions, popup.menu)
-                    popup.menu.findItem(R.id.action_edit).isVisible = item.canEdit
-                    popup.menu.findItem(R.id.action_delete).isVisible = item.canDelete
-                    if (popup.menu is MenuBuilder) {
-                        (popup.menu as MenuBuilder).setOptionalIconsVisible(true)
-                    }
-                    popup.setOnMenuItemClickListener { menuItem ->
-                        when (menuItem.itemId) {
-                            R.id.action_edit -> {
-                                onEditClicked(item)
-                                true
-                            }
-
-                            R.id.action_delete -> {
-                                onDeleteClicked(item)
-                                true
-                            }
-
-                            else -> false
-                        }
-                    }
-                    popup.show()
-                }
-            }
-        }
-
-        companion object {
-            private const val VIEW_TYPE_HEADER = 0
-            private const val VIEW_TYPE_COMMENT = 1
-
-            private val DIFF_CALLBACK = org.ole.planet.myplanet.lite.util.DiffUtils.itemCallback<PostDetailItem>({ oldItem, newItem ->
-                when {
-                    oldItem is PostDetailItem.Header && newItem is PostDetailItem.Header -> oldItem.id == newItem.id
-                    oldItem is PostDetailItem.Comment && newItem is PostDetailItem.Comment -> oldItem.id == newItem.id
-                    else -> false
-                }
-            })
-        }
-    }
-
-    private sealed class PostDetailItem {
-        data class Header(
-            val id: String,
-            val author: String,
-            val username: String?,
-            val hasAvatar: Boolean,
-            val message: String?,
-            val imagePaths: List<String>,
-            val timestamp: Long,
-            val commentCount: Int,
-            val isLoadingComments: Boolean,
-            val canReply: Boolean,
-            val canEdit: Boolean,
-            val canDelete: Boolean,
-            val canShare: Boolean
-        ) : PostDetailItem()
-
-        data class Comment(
-            val id: String,
-            val author: String,
-            val username: String?,
-            val hasAvatar: Boolean,
-            val message: String?,
-            val imagePaths: List<String>,
-            val timestamp: Long,
-            val canEdit: Boolean,
-            val canDelete: Boolean,
-            val document: NewsDocument?
-        ) : PostDetailItem()
-    }
 
     companion object {
         const val EXTRA_POST_ID = "extra_post_id"
@@ -1913,7 +1579,7 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         private const val EXPANDED_REPLY_MIN_LINES = 3
         private const val MAX_HEADING_LEVEL = 6
 
-        private fun formatRelativeTime(context: Context, timestamp: Long): String {
+        internal fun formatRelativeTime(context: Context, timestamp: Long): String {
             val now = System.currentTimeMillis()
             val diffMillis = max(0L, now - timestamp)
             val minutes = diffMillis / MINUTE_MILLIS
@@ -1951,7 +1617,7 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
             }
         }
 
-        private fun buildMetadata(username: String?, relativeTime: String): String {
+        internal fun buildMetadata(username: String?, relativeTime: String): String {
             return if (!username.isNullOrBlank()) {
                 "@${username.trim()} • $relativeTime"
             } else {
